excluded Kernel.php from coverage reports
