+++ src/DataFixtures/AppFixtures.php
@@ -8,7 +8,7 @@ use Doctrine\Bundle\FixturesBundle\Fixture;
 use Doctrine\Persistence\ObjectManager;
 use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
 
-class AppFixtures //extends Fixture
+class AppFixtures extends Fixture
 {
     private $encoder;
 
@@ -19,8 +19,20 @@ class AppFixtures //extends Fixture
 
     public function load(ObjectManager $manager)
     {
-        $manager->persist($this->createAnonymousUser());
-        $manager->persist($this->createOriginalAdmin());
+        $anonymousUser = $this->createAnonymousUser();
+        $manager->persist($anonymousUser);
+
+        $adminUser = $this->createAdminUser();
+        $manager->persist($adminUser);
+
+        $basicUser = $this->createBasicUser();
+        $manager->persist($basicUser);
+
+        $manager->flush();
+
+        $manager->persist($this->createUnlinkedTask($anonymousUser));
+        $manager->persist($this->createAdminTask($adminUser));
+        $manager->persist($this->createBasicTask($basicUser));
 
         $manager->flush();
     }
@@ -35,14 +47,55 @@ class AppFixtures //extends Fixture
             ;
     }
 
-    private function createOriginalAdmin(): User
+    private function createAdminUser(): User
     {
-        $originalAdmin = new User;
-        return $originalAdmin
+        $adminUser = new User;
+        return $adminUser
             ->setEmail('admin@changezmoi.fr')
-            ->setPassword($this->encoder->encodePassword($originalAdmin, 'admin'))
+            ->setPassword($this->encoder->encodePassword($adminUser, 'admin'))
             ->setAdmin(true)
             ->setUsername('administrateur')
+            ;
+    }
+
+    private function createBasicUser()
+    {
+        $basicUser = new User;
+        return $basicUser
+            ->setEmail('basic@changezmoi.fr')
+            ->setPassword($this->encoder->encodePassword($basicUser, 'basic'))
+            ->setAdmin(false)
+            ->setUsername('utilisateur basique')
+            ;
+    }
+
+    private function createUnlinkedTask(User $anonymous): Task
+    {
+        $anonymousTask = new Task;
+        return $anonymousTask
+            ->setTitle('Tache anonyme')
+            ->setContent('Cette tache n\'est attachée à aucun utilisateur particulier')
+            ->setOwner($anonymous)
+            ;
+    }
+
+    private function createAdminTask(User $admin): Task
+    {
+        $adminTask = new Task;
+        return $adminTask
+            ->setTitle('Tache administrateur')
+            ->setContent('Cette tache appartient à l\'administrateur.')
+            ->setOwner($admin)
+            ;
+    }
+
+    private function createBasicTask(User $user): Task
+    {
+        $basicTask = new Task;
+        return $basicTask
+            ->setTitle('Tache classique')
+            ->setContent('Cette tache appartient à un utilisateur normal.')
+            ->setOwner($user)
             ;
     }
 }
